refactor: all done

+++ mysql/45_function.php
@@ -1,5 +1,24 @@
 <?php include "45_db.php"; ?>
 <?php
+function createRows() {
+
+if (isset($_POST['submit'])){
+    global $connection;
+    $username = $_POST['username'];
+    $password = $_POST['password'];
+     
+     $query = "INSERT INTO users(username, password) ";
+     $query .= "VALUES ('$username', '$password')";
+     $result = mysqli_query($connection, $query);
+     if (!$result) {
+         die('Query FAILED' . mysqli_error($connection));
+     } else {
+         echo "Record Create";
+     }
+    }  
+  
+}
+
 function showAllDate () {
 global $connection;
 $query = "SELECT * FROM users";
@@ -16,6 +35,7 @@ while($row = mysqli_fetch_assoc($result)) {
 }
 
 function updateTable(){
+    if (isset($_POST['submit'])){
     global $connection;
     $username = $_POST['username'];
     $password = $_POST['password'];
@@ -24,17 +44,24 @@ function updateTable(){
     $result = mysqli_query($connection, $query);
     if (!$result) {
         die("QUERY FAILED" . mysqli_error($connection));
+    } else {
+        echo "Record Update";
     }
-
+ }
 }
 
 function deleteRows(){
+    if (isset($_POST['submit'])){
     global $connection;
     $id = $_POST['id'];
     $query = "DELETE FROM users WHERE id = $id";
     $result = mysqli_query($connection, $query);
     if (!$result) {
         die("QUERY FAILED" . mysqli_error($connection));
+    }
+    else {
+        echo "Record Deleted";
+    }
     }
 }
 ?>
